refactor: 점수 계산 기능을 `Hand`에 위임
