+++ src/main/java/blackjack/model/Player.java
@@ -7,9 +7,9 @@ public class Player {
     private final String name;
     private final Hand hand;
 
-    public Player(String name) {
+    public Player(String name, Hand hand) {
         this.name = name;
-        this.hand = new Hand();
+        this.hand = hand;
     }
 
     public String getName() {
