+++ src/test/java/blackjack/model/HandTest.java
@@ -10,6 +10,7 @@ import org.junit.jupiter.api.Test;
 class HandTest {
 
     private final Card card = new Card(Rank.ACE, Suit.CLUB);
+    private final AceAdjustPolicy aceAdjustPolicy = new AceAdjustPolicy(new BustPolicyImpl());
 
     @Nested
     @DisplayName("1장 이상의 카드를 받아서 손패에 추가한다.")
@@ -18,7 +19,7 @@ class HandTest {
         @DisplayName("한장의 카드를 받아서 손패에 추가한다.")
         void addOneCard() {
             // given
-            Hand hand = new Hand();
+            Hand hand = new Hand(aceAdjustPolicy);
             List<Card> existCards = hand.getCards();
 
             // when
@@ -30,4 +31,25 @@ class HandTest {
         }
     }
 
+    @Test
+    @DisplayName("카드 점수의 합을 계산한다")
+    void calculateTotalScore() {
+        // given
+        Hand hand = new Hand(aceAdjustPolicy);
+        List<Card> cards = List.of(
+                new Card(Rank.TWO, Suit.CLUB),
+                new Card(Rank.THREE, Suit.CLUB)
+        );
+        hand.addCards(cards);
+
+        int expectedScore = cards.stream()
+                .mapToInt(Card::getValue)
+                .sum();
+
+        // when
+        int actualScore = hand.calculateScore();
+
+        // then
+        assertThat(actualScore).isEqualTo(expectedScore);
+    }
 }
