+++ src/test/java/blackjack/model/PlayerTest.java
@@ -7,6 +7,8 @@ import org.junit.jupiter.api.Test;
 
 class PlayerTest {
 
+    private final Hand hand = new Hand(new AceAdjustPolicy(new BustPolicyImpl()));
+
     @Test
     @DisplayName("본인의 이름을 반환한다.")
     void returnName() {
@@ -14,7 +16,7 @@ class PlayerTest {
         String playerName = "Player Name";
 
         // when
-        Player player = new Player(playerName);
+        Player player = new Player(playerName, hand);
 
         // then
         assertThat(player.getName()).isEqualTo(playerName);
